push

+++ ft_irc/src/Command.cpp
@@ -55,7 +55,7 @@ void Command::parsecmd(std::string& ex)
     params = split(ex);
     std::cout << cmd << std::endl;
     int i = 0;
-    while (i < params.size())
+    while (i < (int)params.size())
         std::cout << params[i++] << std::endl;
 }
